ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-cerebras==0.1.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0